Поправил удаление продукта

+++ wood/views.py
@@ -16,6 +16,7 @@ from django.http import HttpResponse
 
 from django.http import JsonResponse
 from django.shortcuts import render, redirect
+from django.template.loader import render_to_string
 
 from wood.models import *
 from wood_site.settings import STATIC_URL
@@ -92,7 +93,8 @@ def create_product(request, category_id):
 
 def delete_product(request, category_id, product_id):
     product = Product.objects.get(pk=product_id)
-    product.product_image.delete(save=True)
+    if ConcreteProduct.objects.filter(product=product).exists():
+        return error(request, 'В удаляемом изделии существуют конкретные изделия')
     product.delete()
     return redirect('category_content', category_id)
